\% Generated by roxygen2 (4.1.1): do not edit by hand
\% Please edit documentation in R/write.xlsx.R
\name{write.xlsx}
\alias{write.xlsx}
\title{Write to Microsoft Excel XLSX file}
\usage{
write.xlsx(x, xlsx, verbose = FALSE, row.names = FALSE, col.names = TRUE,
  ...)
}
\arguments{
\item{x}{either a \code{data.frame}, or a list of \code{data.frame}s. if a list of
\code{data.frame}s, then each \code{data.frame} will become a worksheet in the resulting
workbook.}

\item{xlsx}{the xlsx file to be created.}

\item{verbose}{logical: print messages?}

\item{row.names}{DIFFERENT TO \code{\link{write.table}}. Can be either \code{TRUE}, \code{FALSE}, \code{NULL},
\code{NA}, or a \code{character(1)}. if \code{TRUE} then row.names will be written out, with a
blank column header, if \code{FALSE}, then no row names will be written out, if
\code{NULL} or \code{NA}, then rownames will be written out with no column name (just
like \code{write.table(row.names=TRUE)} would do) if \code{row.names} is a character
vector of length 1, then write the \code{row.names} out, and use use the supplied
word as the column name for the rownames else error}

\item{col.names}{logical: If \code{TRUE} then write out the column names.}

\item{\dots}{additional arguments passed to \code{\link{write.delim}}}
}
\value{
none. writes an excel 'xlsx' file.
}
\description{
Write a \code{matrix} or \code{data.frame} to a genuine Microsoft Excel XLSX file.
The first row is bold, italics, dashed underline. Columns are centred.
}
\details{
Uses a perl module called \code{Excel::Writer::XLSX}, by John McNamara,
a perl program called \code{tab2xlsx}, based on a program of the same name also by
JM.
This R code borrows the structure from \code{gdata}'s excellent
\code{\link[gdata]{read.xls}} function.
}
\note{
Installation
It's not that widely tested, but this package does include the necessary perl libraries to
write XLSX files. If you're struggling though, install the `Excel::Writer::XLSX` Perl module
via CPAN or similar. For OSX users, this involves:\cr
sudo perl -e shell -MCPAN\cr
install Excel::Writer::XLSX\cr
}
\examples{
f <- tempfile(fileext=".xlsx")
write.xlsx(iris, f)
x <- list(A=iris[1:5, ], B=iris)
write.xlsx(x, f)
}
\author{
Mark Cowley, 2009-01-20. Updated 2016-04-14 to write XLSX files.
}
\seealso{
\code{\link{write.xls}} \code{\link{write.table}} \code{\link{read.delim}} \code{\link[gdata]{read.xls}}
}

