\% Generated by roxygen2: do not edit by hand
\% Please edit documentation in R/writeLines.xlsx.R
\name{writeLines.xlsx}
\alias{writeLines.xlsx}
\title{Write a vector out to an xlsx file.}
\usage{
writeLines.xlsx(x, file, names = TRUE)
}
\arguments{
\item{x}{a vector}

\item{file}{the path to an output file}

\item{names}{logical: if \code{TRUE}, then write out a 2 column table of names then value, 
if \code{FALSE}, then write out just the values in \code{x}.}
}
\value{
nothing.
}
\description{
Write a vector out to an xlsx file. If the vector has names, then optionally
write out a two column xlsx file with names in 1st col, data in 2nd.
See the help for \code{\link{write.xlsx}} for the configuration required.
}
\author{
Mark Cowley, 2009-02-03
}
\seealso{
\code{\link{write.xlsx}}
}

